requirements.txt
adrf==0.1.4
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
asttokens==2.4.1
async-property==0.2.2
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
beartype==0.17.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
cx-Oracle==8.3.0
distro==1.9.0
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
docker==7.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
fast-bitrix24==1.5.16
frozenlist==1.4.1
icontract==2.6.6
idna==3.6
jsonschema==3.2.0
more-itertools==10.2.0
multidict==6.0.5
packaging==23.2
paramiko==3.4.0
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.20.0
python-dotenv==0.21.1
pytz==2024.1
PyYAML==5.3.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
texttable==1.7.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
websocket-client==0.59.0
yarl==1.9.4
